Move structs into modules corresponding to Rust modules
