+++ pineappl_py/tests/test_boc.py
@@ -1,0 +1,7 @@
+import pineappl
+
+
+class TestChannel:
+    def test_init(self):
+        le = pineappl.boc.Channel([(2, 2, 0.5)])
+        assert isinstance(le, pineappl.boc.Channel)
