--- pineappl_py/tests/test_channel.py
@@ -1,7 +1,0 @@
-import pineappl
-
-
-class TestChannel:
-    def test_init(self):
-        le = pineappl.channel.Channel([(2, 2, 0.5)])
-        assert isinstance(le, pineappl.channel.Channel)
