+++ pineappl_py/tests/test_sugrid.py
@@ -11,7 +11,7 @@ class TestSubgridParams:
 
 
 def test_issue_164(pdf):
-    channels = [pineappl.channel.Channel([(1, 2, 1.0)])]
+    channels = [pineappl.boc.Channel([(1, 2, 1.0)])]
     orders = [pineappl.grid.Order(0, 0, 0, 0)]
     params = pineappl.subgrid.SubgridParams()
 
@@ -33,7 +33,7 @@ def test_issue_164(pdf):
 
 class TestSubgrid:
     def fake_grid(self):
-        channels = [pineappl.channel.Channel([(1, 2, 1.0)])]
+        channels = [pineappl.boc.Channel([(1, 2, 1.0)])]
         orders = [pineappl.grid.Order(0, 0, 0, 0)]
         params = pineappl.subgrid.SubgridParams()
         bin_limits = np.array([0.0, 1.0])
